+++ SKPhotoBrowser/SKAnimator.swift
@@ -84,7 +84,10 @@ class SKAnimator: NSObject, SKPhotoBrowserAnimatorDelegate {
             let scrollView = browser.pageDisplayedAtIndex(browser.currentPageIndex) else {
                 
             senderViewForAnimation?.isHidden = false
-            browser.dismissPhotoBrowser(animated: false)
+            browser.dismissPhotoBrowser(animated: false) {
+                self.resizableImageView?.removeFromSuperview()
+                self.backgroundView.removeFromSuperview()
+            }
             return
         }
 
@@ -170,7 +173,7 @@ private extension SKAnimator {
             delay: 0,
             usingSpringWithDamping: animationDamping,
             initialSpringVelocity: 0,
-            options: UIViewAnimationOptions(),
+            options: UIView.AnimationOptions(),
             animations: {
                 browser.showButtons()
                 self.backgroundView.alpha = 1.0
@@ -192,7 +195,7 @@ private extension SKAnimator {
             delay: 0,
             usingSpringWithDamping: animationDamping,
             initialSpringVelocity: 0,
-            options: UIViewAnimationOptions(),
+            options: UIView.AnimationOptions(),
             animations: {
                 self.backgroundView.alpha = 0.0
                 self.resizableImageView?.layer.frame = finalFrame