+++ SKPhotoBrowser/SKButtons.swift
@@ -23,9 +23,10 @@ class SKButton: UIButton {
         }
     }
     fileprivate let size: CGSize = CGSize(width: 44, height: 44)
-    fileprivate let buttonTopOffset: CGFloat = 30
-    fileprivate var margin: CGFloat = SKPhotoBrowserOptions.closeAndDeleteButtonPadding
-
+    fileprivate var marginX: CGFloat = 0
+    fileprivate var marginY: CGFloat = 0
+    fileprivate var extraMarginY: CGFloat = SKMesurement.isPhoneX ? 10 : 0
+    
     func setup(_ imageName: String) {
         backgroundColor = .clear
         imageEdgeInsets = insets
@@ -44,27 +45,43 @@ class SKButton: UIButton {
     func setFrameSize(_ size: CGSize? = nil) {
         guard let size = size else { return }
         
-        let newRect = CGRect(x: margin, y: buttonTopOffset, width: size.width, height: size.height)
+        let newRect = CGRect(x: marginX, y: marginY, width: size.width, height: size.height)
         frame = newRect
         showFrame = newRect
-        hideFrame = CGRect(x: margin, y: -20, width: size.width, height: size.height)
+        hideFrame = CGRect(x: marginX, y: -marginY, width: size.width, height: size.height)
     }
     
     func updateFrame(_ frameSize: CGSize) { }
 }
 
 class SKImageButton: SKButton {
-    
-    fileprivate var leftSidePositionMargin: CGFloat {
-        return super.margin
+    fileprivate var imageName: String { return "" }
+
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
     }
     
-    fileprivate var rightSidePositionMargin: CGFloat {
-        return SKMesurement.screenWidth - super.margin - self.size.width
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setup(imageName)
+        showFrame = CGRect(x: marginX, y: marginY, width: size.width, height: size.height)
+        hideFrame = CGRect(x: marginX, y: -marginY, width: size.width, height: size.height)
     }
-    
-    fileprivate var imageName: String {
-        return ""
+}
+
+class SKCloseButton: SKImageButton {
+    override var imageName: String { return "btn_common_close_wh" }
+    override var marginX: CGFloat {
+        get {
+            return SKPhotoBrowserOptions.swapCloseAndDeleteButtons
+                ? SKMesurement.screenWidth - SKButtonOptions.closeButtonPadding.x - self.size.width
+                : SKButtonOptions.closeButtonPadding.x
+        }
+        set { super.marginX = newValue }
+    }
+    override var marginY: CGFloat {
+        get { return SKButtonOptions.closeButtonPadding.y + extraMarginY }
+        set { super.marginY = newValue }
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -74,37 +91,34 @@ class SKImageButton: SKButton {
     override init(frame: CGRect) {
         super.init(frame: frame)
         setup(imageName)
-        showFrame = CGRect(x: margin, y: buttonTopOffset, width: size.width, height: size.height)
-        hideFrame = CGRect(x: margin, y: -20, width: size.width, height: size.height)
-    }
-}
-
-class SKCloseButton: SKImageButton {
-    override var imageName: String { return "btn_common_close_wh" }
-    override var margin: CGFloat {
-        get {
-            return SKPhotoBrowserOptions.swapCloseAndDeleteButtons
-                ? rightSidePositionMargin
-                : leftSidePositionMargin
-        }
-        set {
-            super.margin = newValue
-        }
+        showFrame = CGRect(x: marginX, y: marginY, width: size.width, height: size.height)
+        hideFrame = CGRect(x: marginX, y: -marginY, width: size.width, height: size.height)
     }
 }
 
 class SKDeleteButton: SKImageButton {
     override var imageName: String { return "btn_common_delete_wh" }
-    override var margin: CGFloat {
-        get { return SKPhotoBrowserOptions.swapCloseAndDeleteButtons
-            ? leftSidePositionMargin
-            : rightSidePositionMargin
+    override var marginX: CGFloat {
+        get {
+            return SKPhotoBrowserOptions.swapCloseAndDeleteButtons
+                ? SKButtonOptions.deleteButtonPadding.x
+                : SKMesurement.screenWidth - SKButtonOptions.deleteButtonPadding.x - self.size.width
         }
-        set { super.margin = newValue }
+        set { super.marginX = newValue }
     }
-    
-    override func updateFrame(_ newScreenSize: CGSize) {
-        showFrame = CGRect(x: newScreenSize.width - size.width, y: buttonTopOffset, width: size.width, height: size.height)
-        hideFrame = CGRect(x: newScreenSize.width - size.width, y: -20, width: size.width, height: size.height)
+    override var marginY: CGFloat {
+        get { return SKButtonOptions.deleteButtonPadding.y + extraMarginY }
+        set { super.marginY = newValue }
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setup(imageName)
+        showFrame = CGRect(x: marginX, y: marginY, width: size.width, height: size.height)
+        hideFrame = CGRect(x: marginX, y: -marginY, width: size.width, height: size.height)
     }
 }