+++ SKPhotoBrowser/SKCache.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 open class SKCache {
-    open static let sharedCache = SKCache()
+    public static let sharedCache = SKCache()
     open var imageCache: SKCacheable
 
     init() {
@@ -38,6 +38,14 @@ open class SKCache {
         }
         
         cache.removeImageForKey(key)
+    }
+    
+    open func removeAllImages() {
+        guard let cache = imageCache as? SKImageCacheable else {
+            return
+        }
+        
+        cache.removeAllImages()
     }
 
     open func imageForRequest(_ request: URLRequest) -> UIImage? {
@@ -77,5 +85,9 @@ class SKDefaultImageCache: SKImageCacheable {
 
     func removeImageForKey(_ key: String) {
         cache.removeObject(forKey: key as AnyObject)
+    }
+    
+    func removeAllImages() {
+        cache.removeAllObjects()
     }
 }