+++ SKPhotoBrowser/SKIndicatorView.swift
@@ -16,7 +16,7 @@ class SKIndicatorView: UIActivityIndicatorView {
     override init(frame: CGRect) {
         super.init(frame: frame)
         center = CGPoint(x: frame.width / 2, y: frame.height / 2)
-        activityIndicatorViewStyle = SKPhotoBrowserOptions.indicatorStyle
+        style = SKPhotoBrowserOptions.indicatorStyle
         color = SKPhotoBrowserOptions.indicatorColor
     }
 }