+++ SKPhotoBrowser/SKPaginationView.swift
@@ -14,6 +14,8 @@ class SKPaginationView: UIView {
     var counterLabel: UILabel?
     var prevButton: UIButton?
     var nextButton: UIButton?
+    private var margin: CGFloat = 100
+    private var extraMargin: CGFloat = SKMesurement.isPhoneX ? 40 : 0
     
     fileprivate weak var browser: SKPhotoBrowser?
     
@@ -26,9 +28,10 @@ class SKPaginationView: UIView {
     }
     
     convenience init(frame: CGRect, browser: SKPhotoBrowser?) {
-        self.init(frame: CGRect(x: 0, y: frame.height - 100, width: frame.width, height: 100))
+        self.init(frame: frame)
+        self.frame = CGRect(x: 0, y: frame.height - margin - extraMargin, width: frame.width, height: 100)
         self.browser = browser
-        
+
         setupApperance()
         setupCounterLabel()
         setupPrevButton()
@@ -52,7 +55,7 @@ class SKPaginationView: UIView {
     }
     
     func updateFrame(frame: CGRect) {
-        self.frame = CGRect(x: 0, y: frame.height - 100, width: frame.width, height: 100)
+        self.frame = CGRect(x: 0, y: frame.height - margin, width: frame.width, height: 100)
     }
     
     func update(_ currentPageIndex: Int) {
@@ -85,6 +88,8 @@ private extension SKPaginationView {
     }
     
     func setupCounterLabel() {
+        guard SKPhotoBrowserOptions.displayCounterLabel else { return }
+        
         let label = UILabel(frame: CGRect(x: 0, y: 0, width: 100, height: 50))
         label.center = CGPoint(x: frame.width / 2, y: frame.height / 2)
         label.textAlignment = .center
@@ -140,7 +145,7 @@ class SKPaginationButton: UIButton {
         
         let image = UIImage(named: "SKPhotoBrowser.bundle/images/\(imageName)",
             in: bundle, compatibleWith: nil) ?? UIImage()
-        setImage(image, for: UIControlState())
+        setImage(image, for: .normal)
     }
 }