+++ SKPhotoBrowser/SKPhotoBrowserDelegate.swift
@@ -74,11 +74,16 @@ import Foundation
     
     /**
      Tells the delegate that the controls view toggled visibility
+     
+     - Parameter browser: reference to the calling SKPhotoBrowser
+     - Parameter hidden: the status of visibility control
      */
     @objc optional func controlsVisibilityToggled(_ browser: SKPhotoBrowser, hidden: Bool)
     
     /**
      Allows  the delegate to create its own caption view
+     
+     - Parameter index: the index of the photo
      */
     @objc optional func captionViewForPhotoAtIndex(index: Int) -> SKCaptionView?
 }