+++ SKPhotoBrowserExample/SKPhotoBrowserExample/FromLocalViewController.swift
@@ -20,15 +20,13 @@ class FromLocalViewController: UIViewController, UICollectionViewDataSource, UIC
         // Static setup
         SKPhotoBrowserOptions.displayAction = true
         SKPhotoBrowserOptions.displayStatusbar = true
+        SKPhotoBrowserOptions.displayCounterLabel = true
+        SKPhotoBrowserOptions.displayBackAndForwardButton = true
 
         setupTestData()
         setupCollectionView()
     }
 
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-    }
-    
     override var prefersStatusBarHidden: Bool {
         return false
     }
@@ -59,11 +57,9 @@ extension FromLocalViewController {
 
 extension FromLocalViewController {
     @objc(collectionView:didSelectItemAtIndexPath:) func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        
         let browser = SKPhotoBrowser(photos: images, initialPageIndex: indexPath.row)
         browser.delegate = self
-//        browser.updateCloseButton(UIImage(named: "image1.jpg")!)
-        
+
         present(browser, animated: true, completion: {})
     }
     
@@ -101,12 +97,16 @@ extension FromLocalViewController {
         // handle dismissing custom actions
     }
     
-    func removePhoto(index: Int, reload: (() -> Void)) {
+    func removePhoto(_ browser: SKPhotoBrowser, index: Int, reload: @escaping (() -> Void)) {
         reload()
     }
-    
+
     func viewForPhoto(_ browser: SKPhotoBrowser, index: Int) -> UIView? {
         return collectionView.cellForItem(at: IndexPath(item: index, section: 0))
+    }
+    
+    func captionViewForPhotoAtIndex(index: Int) -> SKCaptionView? {
+        return nil
     }
 }
 
@@ -137,7 +137,7 @@ class ExampleCollectionViewCell: UICollectionViewCell {
     override func awakeFromNib() {
         super.awakeFromNib()
         exampleImageView.image = nil
-        layer.cornerRadius = 25.0
+//        layer.cornerRadius = 25.0
         layer.masksToBounds = true
     }