+++ SKPhotoBrowserTests/SKCacheTests.swift
@@ -66,4 +66,23 @@ class SKCacheTests: XCTestCase {
         let cachedImage = self.cache.imageForKey(self.key)
         XCTAssertNil(cachedImage)
     }
+    
+    func testDefaultCacheRemoveAllImages() {
+        // given
+        let cache = (self.cache.imageCache as? SKDefaultImageCache)!.cache
+        cache.setObject(self.image, forKey: self.key as AnyObject)
+        
+        let anotherImage = UIImage()
+        let anotherKey = "another_test_image"
+        cache.setObject(anotherImage, forKey: anotherKey as AnyObject)
+        
+        // when
+        self.cache.removeAllImages()
+        
+        // then
+        let cachedImage = self.cache.imageForKey(self.key)
+        let anotherCachedImage = self.cache.imageForKey(anotherKey)
+        XCTAssertNil(cachedImage)
+        XCTAssertNil(anotherCachedImage)
+    }
 }